// src/generated/models/Pm_workingdaysesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_workingdaysespm_month = {
  1: 'January',
  2: 'February',
  3: 'March',
  4: 'April',
  5: 'May',
  6: 'June',
  7: 'July',
  8: 'August',
  9: 'September',
  10: 'October',
  11: 'November',
  12: 'December'
} as const;
export type Pm_workingdaysespm_month = keyof typeof Pm_workingdaysespm_month;
export const Pm_workingdaysesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_workingdaysesstatecode = keyof typeof Pm_workingdaysesstatecode;
export const Pm_workingdaysesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_workingdaysesstatuscode = keyof typeof Pm_workingdaysesstatuscode;

export interface Pm_workingdaysesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  "pm_businessunit@odata.bind"?: string;
  pm_month?: Pm_workingdaysespm_month;
  pm_name: string;
  pm_workingdays?: number;
  pm_workingdaysid: string;
  pm_year?: number;
  statecode: Pm_workingdaysesstatecode;
  statuscode?: Pm_workingdaysesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_workingdayses extends Pm_workingdaysesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_businessunitname?: string;
  pm_monthname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_businessunit?: object;
  _pm_businessunit_value?: string;
}